# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Design Print Production Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # Users
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Quotations
...
  /job-orders:
    get:
      operationId: listJobOrders
      tags: [jobOrders]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: productionLine
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of job orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobOrder"
    post:
      operationId: createJobOrder
      tags: [jobOrders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobOrder"
  /job-orders/{id}:
    get:
      operationId: getJobOrder
      tags: [jobOrders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job order with stages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobOrderDetail"
    patch:
      operationId: updateJobOrder
      tags: [jobOrders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobOrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobOrder"
  /job-orders/{id}/advance-stage:
    post:
      operationId: advanceJobStage
      tags: [jobOrders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StageAdvanceInput"
      responses:
        "200":
          description: Stage advanced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobOrder"

  # Job Stages
  /job-orders/{id}/stages:
    get:
      operationId: listJobStages
      tags: [jobStages]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job stages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobStage"
  /job-stages/{id}:
    patch:
      operationId: updateJobStage
      tags: [jobStages]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobStageUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStage"

  # Machines
...
  /job-orders/{id}/costing:
    get:
      operationId: getJobCosting
      tags: [jobCosting]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job costing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobCosting"
    put:
      operationId: upsertJobCosting
      tags: [jobCosting]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobCostingInput"
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobCosting"

  # QC
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, username, name, role, isActive, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, sales, operations, designer, finalizer, production_manager, machine_operator, qc, warehouse, finance, delivery]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    Customer:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        companyName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        customerType:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    CustomerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        companyName:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        customerType:
          type: string
        source:
          type: string
        notes:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        companyName:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        customerType:
          type: string
        source:
          type: string
        notes:
          type: string
...
    JobOrder:
      type: object
      required: [id, jobNumber, customerId, productType, productionLine, quantity, priority, status, currentStage, createdAt]
      properties:
        id:
          type: integer
        jobNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        salesOrderId:
          type: ["integer", "null"]
        productType:
          type: string
        productionLine:
          type: string
          enum: [offset_printing, silk_screen, plastic_bags, corrugated_carton, flowpack]
        quantity:
          type: integer
        designFileStatus:
          type: string
          enum: [pending, in_review, approved, rejected]
        requiredDeliveryDate:
          type: ["string", "null"]
        priority:
          type: string
          enum: [normal, urgent, very_urgent]
        currentStage:
          type: ["string", "null"]
        assignedDepartment:
          type: ["string", "null"]
        assignedUserId:
          type: ["integer", "null"]
        assignedUserName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, on_hold, completed, delivered, cancelled]
        isDelayed:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    JobOrderDetail:
      type: object
      required: [id, jobNumber, customerId, productType, productionLine, quantity, priority, status, currentStage, createdAt, stages]
      properties:
        id:
          type: integer
        jobNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        salesOrderId:
          type: ["integer", "null"]
        productType:
          type: string
        productionLine:
          type: string
        quantity:
          type: integer
        designFileStatus:
          type: string
        requiredDeliveryDate:
          type: ["string", "null"]
        priority:
          type: string
        currentStage:
          type: ["string", "null"]
        assignedDepartment:
          type: ["string", "null"]
        assignedUserId:
          type: ["integer", "null"]
        assignedUserName:
          type: ["string", "null"]
        status:
          type: string
        isDelayed:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        stages:
          type: array
          items:
            $ref: "#/components/schemas/JobStage"
    JobOrderInput:
      type: object
      required: [customerId, productType, productionLine, quantity, priority]
      properties:
        customerId:
          type: integer
        salesOrderId:
          type: integer
        productType:
          type: string
        productionLine:
          type: string
          enum: [offset_printing, silk_screen, plastic_bags, corrugated_carton, flowpack]
        quantity:
          type: integer
        designFileStatus:
          type: string
        requiredDeliveryDate:
          type: string
        priority:
          type: string
          enum: [normal, urgent, very_urgent]
        assignedDepartment:
          type: string
        assignedUserId:
          type: integer
        notes:
          type: string
    JobOrderUpdate:
      type: object
      properties:
        productType:
          type: string
        quantity:
          type: integer
        designFileStatus:
          type: string
        requiredDeliveryDate:
          type: string
        priority:
          type: string
        currentStage:
          type: string
        assignedDepartment:
          type: string
        assignedUserId:
          type: integer
        status:
          type: string
        notes:
          type: string
    StageAdvanceInput:
      type: object
      required: [stageName]
      properties:
        stageName:
          type: string
        notes:
          type: string
        delayReason:
          type: string
    JobStage:
      type: object
      required: [id, jobOrderId, stageName, stageOrder, status]
      properties:
        id:
          type: integer
        jobOrderId:
          type: integer
        stageName:
          type: string
        stageOrder:
          type: integer
        status:
          type: string
          enum: [waiting, in_progress, done, skipped, on_hold]
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        responsibleUserId:
          type: ["integer", "null"]
        responsibleUserName:
          type: ["string", "null"]
        machineId:
          type: ["integer", "null"]
        machineName:
          type: ["string", "null"]
        delayReason:
          type: ["string", "null"]
        wasteQuantity:
          type: ["number", "null"]
        actualOutput:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    JobStageUpdate:
      type: object
      properties:
        status:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        responsibleUserId:
          type: integer
        machineId:
          type: integer
        delayReason:
          type: string
        wasteQuantity:
          type: number
        actualOutput:
          type: number
        notes:
          type: string
...
    JobCosting:
      type: object
      required: [id, jobOrderId]
      properties:
        id:
          type: integer
        jobOrderId:
          type: integer
        materialCost:
          type: ["number", "null"]
        machineCost:
          type: ["number", "null"]
        laborCost:
          type: ["number", "null"]
        outsourcingCost:
          type: ["number", "null"]
        wasteCost:
          type: ["number", "null"]
        totalEstimatedCost:
          type: ["number", "null"]
        actualCost:
          type: ["number", "null"]
        sellingPrice:
          type: ["number", "null"]
        grossProfit:
          type: ["number", "null"]
        grossMarginPercent:
          type: ["number", "null"]
    JobCostingInput:
      type: object
      properties:
        materialCost:
          type: number
        machineCost:
          type: number
        laborCost:
          type: number
        outsourcingCost:
          type: number
        wasteCost:
          type: number
        totalEstimatedCost:
          type: number
        actualCost:
          type: number
        sellingPrice:
          type: number
...