# lib/api-zod/src/generated/types/jobOrderUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Design Print Production Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface JobOrderUpdate {
  productType?: string;
  quantity?: number;
  designFileStatus?: string;
  requiredDeliveryDate?: string;
  priority?: string;
  currentStage?: string;
  assignedDepartment?: string;
  assignedUserId?: number;
  status?: string;
  notes?: string;
}